Incorporated collisions for entitied
I based the collisions off my 3d alpha wold game engine, minus a channel. There is an interface called a collision. I have created a class called "collisionBox" that implements "collision". I was working on a collision sphere, but ran into trouble and decided it wasn't necessary at this point.

I'm thinking about scrapping the tile based terrain system. Everything will be on a grid, but be entity based with an artistic background.